requirements.txt
bbd-libs-pybus==1.2.3
bbd.libs.python==4.0.2
bbdcommon.communications.systemvipc==1.1.15 ;sys_platform == 'linux'
bbd-services-contentstore==2.0.1b8
sataxi-finance-messaging==1.0.0.312
psycopg2-binary==2.9.3
apispec==5.1.1
requests==2.25.1
pysmb==1.2.8
xlsxwriter==3.0.3
pandas==1.3.5
importlib-metadata<5.0
pdfkit==1.0.0
typing-extensions==4.5.0
colorama<0.4.5
psycopg2==2.9.5
